Moved classes and methods that manipulate the parse tree to its own file.
